// Frontend/src/utils/enum.ts
export enum RoutePaths {
  Login = "/login",
  Register = "/register",
  Home = "/",
  CustomerPage = "/customer",
  Account = "/Account",
  AccountList = "/AccountList",
  FolderList = "/FolderList",
  Folders = "/folders",
  UploadFile = "/uploadfile",
  ForgotPassword = "/forgotpassword"
}